[SASS-4940] PIP: new SimpleRedirectService
