+++ app/controllers/pensions/paymentsIntoPensions/ReliefAtSourcePensionsController.scala
@@ -24,7 +24,7 @@ import models.mongo.PensionsCYAModel
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import services.PensionSessionService
-import services.RedirectService.isFinishedCheck
+import services.SimpleRedirectService.isFinishedCheck
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
 import utils.Clock
 import views.html.pensions.paymentsIntoPensions.ReliefAtSourcePensionsView
@@ -65,30 +65,30 @@ class ReliefAtSourcePensionsController @Inject()(authAction: AuthorisedAction,
         pensionSessionService.getPensionSessionData(taxYear, request.user).flatMap {
           case Right(optData) =>
 
-          val pensionsCya = optData.map(_.pensions).getOrElse(PensionsCYAModel.emptyModels)
-          val viewModel = pensionsCya.paymentsIntoPension
+            val pensionsCya = optData.map(_.pensions).getOrElse(PensionsCYAModel.emptyModels)
+            val viewModel = pensionsCya.paymentsIntoPension
 
-          val updatedCyaModel = {
-            pensionsCya.copy(
-              paymentsIntoPension = viewModel.copy(
-                rasPensionPaymentQuestion = Some(yesNo),
-                totalRASPaymentsAndTaxRelief = if (yesNo) viewModel.totalRASPaymentsAndTaxRelief else None,
-                oneOffRasPaymentPlusTaxReliefQuestion = if (yesNo) viewModel.oneOffRasPaymentPlusTaxReliefQuestion else None,
-                totalOneOffRasPaymentPlusTaxRelief = if (yesNo) viewModel.totalOneOffRasPaymentPlusTaxRelief else None,
-                totalPaymentsIntoRASQuestion = if (yesNo) viewModel.totalPaymentsIntoRASQuestion else None
+            val updatedCyaModel = {
+              pensionsCya.copy(
+                paymentsIntoPension = viewModel.copy(
+                  rasPensionPaymentQuestion = Some(yesNo),
+                  totalRASPaymentsAndTaxRelief = if (yesNo) viewModel.totalRASPaymentsAndTaxRelief else None,
+                  oneOffRasPaymentPlusTaxReliefQuestion = if (yesNo) viewModel.oneOffRasPaymentPlusTaxReliefQuestion else None,
+                  totalOneOffRasPaymentPlusTaxRelief = if (yesNo) viewModel.totalOneOffRasPaymentPlusTaxRelief else None,
+                  totalPaymentsIntoRASQuestion = if (yesNo) viewModel.totalPaymentsIntoRASQuestion else None
+                )
               )
-            )
-          }
-          val redirectLocation = if (yesNo) {
-            ReliefAtSourcePaymentsAndTaxReliefAmountController.show(taxYear)
-          } else {
-            PensionsTaxReliefNotClaimedController.show(taxYear)
-          }
+            }
+            val redirectLocation = if (yesNo) {
+              ReliefAtSourcePaymentsAndTaxReliefAmountController.show(taxYear)
+            } else {
+              PensionsTaxReliefNotClaimedController.show(taxYear)
+            }
 
-          pensionSessionService.createOrUpdateSessionData(request.user,
-            updatedCyaModel, taxYear, optData.exists(_.isPriorSubmission))(errorHandler.internalServerError()) {
+            pensionSessionService.createOrUpdateSessionData(request.user,
+              updatedCyaModel, taxYear, optData.exists(_.isPriorSubmission))(errorHandler.internalServerError()) {
               isFinishedCheck(updatedCyaModel, taxYear, redirectLocation)
-          }
+            }
         }
       }
     )
